# cmd/web-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "SaaS Example API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Find returns the existing users in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter string, example: name = 'Company Name' and email = 'gabi.may@geeksinthewoods.com'",
                        "name": "where",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order columns separated by comma, example: created_at desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit, example: 10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset, example: 20",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Included Archived, example: false",
                        "name": "include-archived",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.UserResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Create inserts a new user into the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create new user.",
                "parameters": [
                    {
                        "description": "User details",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/user.UserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Update updates the specified user in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user by ID",
                "parameters": [
                    {
                        "description": "Update fields",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/user.UserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/password": {
            "patch": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Update updates the password for a specified user in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user password by ID",
                "parameters": [
                    {
                        "description": "Update fields",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/user.UserUpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "user.UserCreateRequest": {
            "type": "object",
            "required": [
                "branch_id",
                "email",
                "first_name",
                "last_name",
                "password",
                "password_confirm"
            ],
            "properties": {
                "branch_id": {
                    "type": "string",
                    "example": "717cbfd4-b228-48f6-92bc-cc054a4e13f6"
                },
                "email": {
                    "type": "string",
                    "example": "gabi@geeksinthewoods.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Gabi"
                },
                "last_name": {
                    "type": "string",
                    "example": "May"
                },
                "password": {
                    "type": "string",
                    "example": "SecretString"
                },
                "password_confirm": {
                    "type": "string",
                    "example": "SecretString"
                },
                "timezone": {
                    "type": "string",
                    "example": "America/Anchorage"
                }
            }
        },
        "user.UserResponse": {
            "type": "object",
            "properties": {
                "archived_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "created_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "email": {
                    "type": "string",
                    "example": "gabi@geeksinthewoods.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Gabi"
                },
                "gravatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "d69bdef7-173f-4d29-b52c-3edc60baf6a2"
                },
                "last_name": {
                    "type": "string",
                    "example": "May"
                },
                "name": {
                    "type": "string",
                    "example": "Gabi"
                },
                "timezone": {
                    "type": "string",
                    "example": "America/Anchorage"
                },
                "updated_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                }
            }
        },
...
        "weberror.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/weberror.FieldError"
                    }
                },
                "stack_trace": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/v1/oauth/token",
            "scopes": {
                "admin": " Grants administrative privileges with role of admin.",
                "user": " Grants basic privileges with role of user."
            }
        }
    }
}